Refactor : 좋아요 관련 기능 수정
